+++ app/src/main/java/com/example/analyze_credit_system_mobile/view/activity/fragments/CreateCreditDateFragment.kt
@@ -100,7 +100,7 @@ class CreateCreditDateFragment : Fragment() {
                   }
                  else{
                      binding.progresseButton.isClickable = true
-                     val valueInstallment= creditViewModel.getInstallments(number,args.creditValue.toDouble())
+                     val valueInstallment= creditViewModel.getInstallments(number,args.creditValue.toBigDecimal())
                      binding.txvCoutValueInstallment.text= "x R$valueInstallment por mês"
                  }
              }
